Movendo patched_startup para asgard.app
Acabou ficando perdido no asgard.api.agents

Esse código existe por uma falta de funcionalidade no asynworker.
O asyncworker até possui uma infra-estrutura de `startup()` e
`shutdown()` mas ele atualmente faz isso **após** o "listen" da app
http. E o aiohttp não permite mexer na app após esse ponto.

O que esse patched_startup faz é configurar um CORS (ainda inseguro)
para todas as rotas. E precisamos fazer isso antes do "listen".

Quando o asyncworker tiver esse processo de startup divido em dois (pre
listen e pos-listen) poderemos remover esse código.
